Development

+++ app/Http/Controllers/Admin/DepartmentController.php
@@ -7,18 +7,29 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Models\Department;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class DepartmentController extends Controller
 {
     
     public function index()
     { 
+        if(issetAccess(Auth::user()->user_role_id)->location['read'] == false) // Unless the user has access
+        {
+            return redirect('login')->with('exception',trans('app.you_are_not_authorized'));
+        }
+
         $departments = Department::where('company_id', auth()->user()->company_id)->get();
         return view('backend.admin.department.list', compact('departments'));
     }
 
     public function showForm()
     {
+        if(issetAccess(Auth::user()->user_role_id)->location['write'] == false) // Unless the user has access
+        {
+            return redirect('login')->with('exception',trans('app.you_are_not_authorized'));
+        }
+
         $keyList = $this->keyList();
         return view('backend.admin.department.form', compact('keyList'));
     }
@@ -27,6 +38,11 @@ class DepartmentController extends Controller
      * =============================================*/
     public function create(Request $request)
     { 
+        if(issetAccess(Auth::user()->user_role_id)->location['write'] == false)
+        {
+            return redirect('login')->with('exception',trans('app.you_are_not_authorized'));
+        }
+
         @date_default_timezone_set(session('app.timezone'));
 
         $companyId = Auth::user()->company_id;
@@ -45,20 +61,20 @@ class DepartmentController extends Controller
         ));
 
         if ($validator->fails()) {
-            return redirect('admin/department/create')
+            return redirect('admin/location/create')
                 ->withErrors($validator)
                 ->withInput();
         } else {
 
             if(Department::where(['company_id' => $companyId, 'name' => $request->name])->exists())
             {
-                return redirect('admin/department/create')
+                return redirect('admin/location/create')
                 ->withErrors(['name' => 'Department is already exists'])
                 ->withInput();
             }
             else if(Department::where(['company_id' => $companyId, 'key' => $request->key])->exists())
             {
-                return redirect('admin/department/create')
+                return redirect('admin/location/create')
                 ->withErrors(['key' => 'This key is already exists'])
                 ->withInput();
             }
@@ -87,6 +103,11 @@ class DepartmentController extends Controller
  
     public function showEditForm($id = null)
     {
+        if(issetAccess(Auth::user()->user_role_id)->location['write'] == false) // Unless the user has access
+        {
+            return redirect('login')->with('exception',trans('app.you_are_not_authorized'));
+        }
+
         $keyList = $this->keyList();
         $department = Department::where('id', $id)->first();
         return view('backend.admin.department.edit', compact('department', 'keyList'));
@@ -95,6 +116,11 @@ class DepartmentController extends Controller
 
     public function update(Request $request)
     { 
+        if(issetAccess(Auth::user()->user_role_id)->location['write'] == false)
+        {
+            return redirect('login')->with('exception',trans('app.you_are_not_authorized'));
+        }
+
         @date_default_timezone_set(session('app.timezone'));
 
         $companyId = Auth::user()->company_id;
@@ -113,20 +139,20 @@ class DepartmentController extends Controller
         ));
 
         if ($validator->fails()) {
-            return redirect('admin/department/edit/'.$request->id)
+            return redirect('admin/location/edit/'.$request->id)
                         ->withErrors($validator)
                         ->withInput();
         } else {
 
             if(Department::where('id', "!=", $request->id)->where(['company_id' => $companyId, 'name' => $request->name])->exists())
             {
-                return redirect('admin/department/create')
+                return redirect('admin/location/create')
                 ->withErrors(['name' => 'Department is already exists'])
                 ->withInput();
             }
             else if(Department::where('id', "!=", $request->id)->where(['company_id' => $companyId, 'key' => $request->key])->exists())
             {
-                return redirect('admin/department/create')
+                return redirect('admin/location/create')
                 ->withErrors(['key' => 'This key is already exists'])
                 ->withInput();
             }
@@ -153,8 +179,14 @@ class DepartmentController extends Controller
  
     public function delete($id = null)
     {
+        if(issetAccess(Auth::user()->user_role_id)->location['write'] == false) // Unless the user has access
+        {
+            return redirect('login')->with('exception',trans('app.you_are_not_authorized'));
+        }
+
         $delete = Department::where('id', $id)->delete();
-        return redirect('admin/department')->with('message', trans('app.delete_successfully'));
+        DB::table('token_setting')->where('department_id', $id)->delete();
+        return redirect('admin/location')->with('message', trans('app.delete_successfully'));
     } 
  
     public function keyList()
